implement domU monitoring and auto-restart

+++ src/server/lib/perl5/OpenApp/VMMgmt.pm
@@ -15,11 +15,71 @@ our $LIBVIRT_DIR = '/var/oa/libvirt';
 
 my $STATUS_DIR = '/var/run/vmstatus';
 my $HWMON_FILE = '/var/run/vmstatus.hw';
+my $STATUS_LOCK = "$STATUS_DIR/.lock";
 
 my $OA_VERSION_FILE = '/opt/vyatta/etc/version';
 my $OA_GRUB_CFG = '/live/image/boot/grub/grub.cfg';
 
 ### "static" functions
+sub _lockStatus {
+  system("sudo mkdir $STATUS_DIR");
+  while (1) {
+    system("sudo mkdir $STATUS_LOCK >&/dev/null");
+    last if (!($? >> 8));
+    sleep 1;
+  }
+}
+
+sub _unlockStatus {
+  system("sudo rmdir $STATUS_LOCK");
+}
+
+my %status_hash = (
+  'starting' => 1,
+  'stopped' => 1,
+  'restart' => 1,
+  'restart-failed' => 1,
+  'out' => 1
+);
+
+sub _isValidStatus {
+  my ($st) = @_;
+  return 0 if (!defined($st));
+  return (defined($status_hash{$st}));
+}
+
+sub _statusSet {
+  my ($status, $id) = @_;
+  return if (!defined($status) || !defined($id));
+  return if (!_isValidStatus($status));
+  return if (! -f "$STATUS_DIR/$id");
+  system("sudo touch '$STATUS_DIR/.$status.$id'");
+}
+
+sub _statusRm {
+  my ($status, $id) = @_;
+  return if (!defined($status) || !defined($id));
+  return if (!_isValidStatus($status));
+  return if (! -f "$STATUS_DIR/$id");
+  system("sudo rm -f '$STATUS_DIR/.$status.$id'");
+}
+
+sub _isStatus {
+  my ($status, $id) = @_;
+  return 0 if (!defined($status) || !defined($id));
+  return 0 if (!_isValidStatus($status));
+  return 0 if (! -f "$STATUS_DIR/$id");
+  return (-f "$STATUS_DIR/.$status.$id");
+}
+
+sub _statusTime {
+  my ($status, $id) = @_;
+  return undef if (!defined($status) || !defined($id));
+  return undef if (!_isValidStatus($status));
+  return undef if (! -f "$STATUS_DIR/.$status.$id");
+  return ((stat("$STATUS_DIR/.$status.$id"))[9]);
+}
+
 sub getVMList {
   my $dd = undef;
   opendir($dd, "$META_DIR") or return;
@@ -28,14 +88,70 @@ sub getVMList {
   return @v;
 }
 
+sub _getVMStartingWait {
+  # TODO get it from config
+  return 300;
+}
+
+sub _getVMStatusTimeout {
+  # TODO get it from config
+  return 180;
+}
+
+sub readStatus {
+  my ($id) = @_;
+  return () if (! -r "$STATUS_DIR/$id");
+  my $fd = undef;
+  open($fd, '<', "$STATUS_DIR/$id") or return ();
+  my $data = <$fd>;
+  close($fd);
+  chomp($data);
+  my ($st, $cpu, $dall, $dfree, $mall, $mfree, $upd, $crit)
+    = split(/ /, $data, 8);
+  return ($st, $cpu, $dall, $dfree, $mall, $mfree, $upd, $crit);
+}
+
 sub updateStatus {
   my ($id, $st, $cpu, $dall, $dfree, $mall, $mfree, $upd, $crit) = @_;
-  my $fd = undef;
-  # make sure the directory exists
-  mkdir($STATUS_DIR);
-  open($fd, '>', "$STATUS_DIR/$id") or return;
-  print $fd "$st $cpu $dall $dfree $mall $mfree $upd $crit\n";
-  close($fd);
+  _lockStatus();
+  my ($ost) = readStatus($id); # old state
+  while (1) {
+    if ("$ost" ne 'up' && "$st" eq 'up') {
+      # transition to up
+      _statusRm('starting', $id);
+      _statusRm('out', $id);
+      _statusRm('restart', $id);
+      _statusRm('restart-failed', $id);
+      # new status will be updated below
+    } elsif (_isStatus('starting', $id)) {
+      # it has been starting
+      my $start_wait = _getVMStartingWait();
+      my $start_time = _statusTime('starting', $id);
+      my $cur_time = time();
+      if (($cur_time - $start_time) <= $start_wait) {
+        # still starting. not updating status.
+        last;
+      }
+      # starting should be done by now. but it's not up yet.
+      # mark it as out.
+      _statusRm('starting', $id);
+      _statusSet('out', $id);
+      # new status will be updated below
+    }
+    
+    if (!_isStatus('out', $id) && "$ost" eq 'up' && "$st" ne 'up') {
+      # transition from up. mark it as out.
+      _statusSet('out', $id);
+    }
+
+    my $fd = undef;
+    if (open($fd, '>', "$STATUS_DIR/$id")) {
+      print $fd "$st $cpu $dall $dfree $mall $mfree $upd $crit\n";
+      close($fd);
+    }
+    last;
+  }
+  _unlockStatus();
 }
 
 sub updateHwMon {
@@ -158,6 +274,22 @@ sub _getLibvirtCfg {
 
 sub startVM {
   my ($id) = @_;
+
+  _lockStatus();
+  if (_isStatus('starting', $id)) {
+    # already starting. do nothing.
+    _unlockStatus();
+    return;
+  }
+  
+  _statusSet('starting', $id);
+  _statusRm('stopped', $id);
+  _statusRm('out', $id);
+  
+  # change status to unknown
+  system("sudo sh -c 'echo \"unknown 0 0 0 0 0  \" >\"$STATUS_DIR/$id\"'");
+  _unlockStatus();
+
   my $lv_cfg = _getLibvirtCfg($id);
   system("sudo virsh -c xen:/// create $lv_cfg");
 }
@@ -174,9 +306,76 @@ sub _waitVmShutOff {
 
 sub shutdownVM {
   my ($id) = @_;
+  
+  _lockStatus();
+  my $starting = _isStatus('starting', $id);
+  _unlockStatus();
+  return if ($starting); # it's starting. do nothing.
+
   system("sudo virsh -c xen:/// shutdown $id");
   _waitVmShutOff($id);
   system("sudo virsh -c xen:/// destroy $id");
+  
+  _lockStatus();
+  _statusSet('stopped', $id);
+  _unlockStatus();
+}
+
+sub needRestart {
+  my ($id) = @_;
+
+  my $ret = 0;
+  _lockStatus();
+  while (1) {
+    # XXX special case:
+    # if this is not UTM, and UTM is out, don't do auto-restart.
+    # (can't get a domU's status if UTM is out or starting.)
+    last if ("$id" ne 'utm' && (_isStatus('out', 'utm')
+                                || _isStatus('starting', 'utm')));
+
+    last if (!_isStatus('out', $id));
+
+    my $out_time = _statusTime('out', $id);
+    last if (!defined($out_time));
+
+    my $timeout = _getVMStatusTimeout();
+    my $cur_time = time();
+    last if (($cur_time - $out_time) <= $timeout);
+
+    # don't restart if already starting
+    last if (_isStatus('starting', $id));
+  
+    # don't restart if stopped
+    last if (_isStatus('stopped', $id));
+
+    if (_isStatus('restart', $id)) {
+      # already started once and never reached 'up'.
+      if (!_isStatus('restart-failed', $id)) {
+        # TODO log message
+        _statusSet('restart-failed', $id);
+      }
+      last;
+    }
+
+    $ret = 1;
+    last;
+  }
+  _unlockStatus();
+  return $ret;
+}
+
+sub tryAutoRestart {
+  my ($id) = @_;
+
+  _lockStatus();
+  _statusSet('restart', $id);
+  _statusRm('starting', $id);
+  _statusRm('stopped', $id);
+  _statusRm('out', $id);
+  _unlockStatus();
+
+  shutdownVM($id);
+  startVM($id);
 }
 
 ### data
@@ -246,17 +445,10 @@ sub _setupMeta {
 
 sub refreshStatus {
   my ($self) = @_;
-  my $id = $self->{_vmId};
-  if (! -r "$STATUS_DIR/$id") {
-    return;
-  }
-  my $fd = undef;
-  open($fd, '<', "$STATUS_DIR/$id") or return;
-  my $data = <$fd>;
-  close($fd);
-  chomp($data);
+  _lockStatus();
   my ($st, $cpu, $dall, $dfree, $mall, $mfree, $upd, $crit)
-    = split(/ /, $data, 8);
+    = readStatus($self->{_vmId});
+  _unlockStatus();
   $self->{_vmState} = $st;
   $self->{_vmCpu} = $cpu;
   $self->{_vmDiskAll} = $dall;
